+++ build_export_templates.sh
@@ -1,0 +1,20 @@
+#!/bin/bash
+
+GREEN='\033[32m';
+NOCOLOR='\033[0m';
+
+echo -e "${GREEN}Cloning Godot Engine source code${NOCOLOR}";
+git clone https://github.com/godotengine/godot;
+(cd godot && git checkout 4.3);
+
+echo -e "${GREEN}Copying necessary build files${NOCOLOR}";
+cp -v build/custom.py build/ljpm.build build/build_custom_templates.sh godot/;
+
+echo -e "${GREEN}Running docker container${NOCOLOR}";
+docker run -it -v ./godot:/godot registry.gitlab.com/skeletonek/ljpm/slb_godot_build:4.3.0 /godot/build_custom_templates.sh;
+
+echo -e "${GREEN}Copying export templates${NOCOLOR}";
+cp -v godot/bin/* export_templates/;
+
+echo -e "${GREEN}Deleting Godot Engine source code${NOCOLOR}";
+sudo rm -rfv godot/;